refactor : User 클래스 추가 및 Dealer, Player 클래스가 User 클래스를 상속 받도록 변경
