+++ src/main/java/domain/user/Dealer.java
@@ -1,20 +1,8 @@
 package domain.user;
 
-import domain.card.Card;
-import domain.card.Cards;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class Dealer {
-
-    private final List<Card> cards = new ArrayList<>();
+public class Dealer extends User{
 
     public Dealer() {
-    }
-
-    public void drawCard() {
-        cards.add(Cards.removeCardOnTop());
     }
 
 }
