+++ src/main/java/domain/user/Player.java
@@ -1,25 +1,13 @@
 package domain.user;
 
-import domain.card.Card;
-import domain.card.Cards;
+public class Player extends User {
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-
-public class Player {
-    
     private final PlayerName name;
     private final BettingMoney bettingMoney;
-    private final List<Card> cards = new ArrayList<>();
 
     public Player(PlayerName name, BettingMoney bettingMoney) {
         this.name = name;
         this.bettingMoney = bettingMoney;
-    }
-
-    public void drawCard() {
-        cards.add(Cards.removeCardOnTop());
     }
 
     public PlayerName getName() {
@@ -28,10 +16,6 @@ public class Player {
 
     public BettingMoney getBettingMoney() {
         return bettingMoney;
-    }
-
-    public List<Card> getCards() {
-        return Collections.unmodifiableList(cards);
     }
 
 }
